Fix tags of flexible tests + other improvements
- Tidy up some test files and move to more appropriate folders
- Remove deprecated calls
- New tag `ops` for selecting the ops flexible tests
